+++ pkg/transportrequest/cts/upload.go
@@ -1,4 +1,4 @@
-package transportrequest
+package cts
 
 import (
 	"fmt"
